requirements.txt
asgiref==3.9.1
boto3==1.40.49
botocore==1.40.49
cffi==2.0.0
cryptography==46.0.1
Django==5.2.6
django-csp==4.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-storages==1.14.6
django-taggit==6.1.0
djangorestframework==3.16.1
gunicorn==23.0.0
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.11.0
